Move the timing failed path from assembly to the kernel timer function.
The previous decision was to save time on register restore when timing
failed, but complicated the interface. Now prefer a simpler interface,
and since timeout should be a rare event, it's ok to waste a little bit
more time, and we save time each time when we exit the kernel.